fix(android): four bugs surfaced by first real-device install
Found by manually building the AAR + APK and running the debug build
on a Samsung Galaxy S25 Ultra (Android 16). The orchestrator's
"verified by inspection" annotations for the Android lane never
compiled Kotlin, so these regressions slipped through S-014 / S-016 /
S-018 / S-015 and only surfaced when the app actually launched.

1. UDL byte-type mismatch. crates/syauth-mobile/src/mobile.udl declared
   the four byte-carrying parameters as `sequence<u8>`. UniFFI 0.29
   maps that to `List<kotlin.UByte>` on the Kotlin side, but every app
   call site (UniffiOobCalculator, UniffiWireSigner,
   SyauthCompanionService.UniffiChallengeVerifier) passes a plain
   `ByteArray`. Seven `Type mismatch: inferred type is List<UByte> but
   ByteArray was expected` errors at compile time. Switching to the
   `bytes` UDL type keeps the Rust signatures as `Vec<u8>` (no Rust
   change) and emits Kotlin signatures of `ByteArray` everywhere.

2. SyauthCompanionService key-type mismatch. The `controllers` map was
   declared as `ConcurrentHashMap<Long, GattServerController>`, but
   `AssociationInfo.getId()` returns a primitive `int` on Android
   31-36 (which surfaces as Kotlin `Int`, not `Long`). The compile
   errors are `Type mismatch: inferred type is Int but Long was
   expected` plus an inference failure at `controllers.remove(...)`.
   Fixed by declaring the map as `<Int, GattServerController>`.

3. cdylib_name mismatch surfaced at runtime as
   `UnsatisfiedLinkError: dlopen failed: library
   "libuniffi_syauth_mobile.so" not found`. UniFFI 0.29's CLI bindgen
   defaults `findLibraryName(...)` to `uniffi_<namespace>`, but the
   cdylib that `cargo build` produces follows the crate's
   `[lib] name = "syauth_mobile"` and so is `libsyauth_mobile.so`.
   Added `crates/syauth-mobile/uniffi.toml` with `[bindings.kotlin]
   cdylib_name = "syauth_mobile"`; the generator now emits
   `return "syauth_mobile"` in findLibraryName, which Android resolves
   to `libsyauth_mobile.so` cleanly.

4. Gradle wrapper distribution URL pointed at
   `gradle-9.5.0-milestone-3-bin.zip`, which is incompatible with
   AGP 8.2.2 (needs Gradle 8.0-8.10). Pinned to `gradle-8.7-bin.zip`
   (the version prrr-android successfully uses).

Verified end-to-end on real hardware:
  - Launch -> OobScreen renders
    "OOB: <four emoji-word OOB code>" computed via HKDF-SHA256 across
    the UniFFI -> JNA -> Kotlin -> Compose pipeline.
  - Tap "Pair" -> PairingState.Idle ("Pair with computer" CTA).
  - Tap CTA -> PairingState.Scanning (spinner + Cancel).
  - Tap Cancel -> PairingState.Idle.
  - System Back -> back to home; OOB + Pair both visible.
  - Zero AndroidRuntime FATAL or UnsatisfiedLinkError across the run.

The Pair -> LescNegotiating -> Bonded path still needs the S-018 real
BT backend + a paired desktop; the Approve flow still needs S-017's
BiometricPrompt + a challenge frame. Both remain "verified by
inspection" on this dev box; tested manually on first hardware sweep.

+++ syauth-android/app/src/main/kotlin/com/sy/syauth/android/bg/SyauthCompanionService.kt
@@ -124,7 +124,10 @@ public class SyauthCompanionService : CompanionDeviceService() {
      * bonded peers in v0.2; for v0.1 only one slot is used at a
      * time but the data structure scales without refactor.
      */
-    private val controllers: ConcurrentHashMap<Long, GattServerController> =
+    // Key type is `Int` (not `Long`): `AssociationInfo.getId()` on
+    // Android API 31-36 returns a primitive `int`. The Kotlin type
+    // mapping is `Int`, so the map key follows.
+    private val controllers: ConcurrentHashMap<Int, GattServerController> =
         ConcurrentHashMap()
 
     override fun onDeviceAppeared(associationInfo: AssociationInfo) {
